refactor: address review feedback
refactor: direct import

refactor: type styled

test: add icon visibility checks

refactor: inline input change gen logic
