+++ src/utils/create-event.js
@@ -7,8 +7,9 @@ LICENSE file in the root directory of this source tree.
 // @flow
 /* global window document */
 
+/** A safe way to create event objects down to IE11 */
 export default function createEvent(eventName: string) {
-  var event;
+  let event;
   if (typeof window.Event === 'function') {
     event = new window.Event(eventName, {bubbles: true, cancelable: true});
   } else {
